+++ index.html
@@ -12,7 +12,7 @@
     <li><h3><a href="examples">Examples</a></h3></li>
     <li><h3><a href="examples/editor/index.html">Interactive editor</a></h3></li>
     <li><h3><a href="https://github.com/CartoDB/carto-vl">GitHub repository</a><h3></li>
-    <li><h3><a href="debug/bq-mvt/index.html">Big Query MVT</a></h3></li>
+    <li><h3><a href="debug/bq-tileset/index.html">Big Query Tileset</a></h3></li>
   </ul>
 </body>
 </html>
