+++ src/index.js
@@ -45,7 +45,7 @@ import * as expressions from './renderer/viz/expressions';
 import Dataset from './sources/Dataset';
 import GeoJSON from './sources/GeoJSON';
 import MVT from './sources/MVT';
-import BQMVT from './sources/bigquery/BQMVT';
+import BigQueryTileset from './sources/bigquery/BQTileset';
 import SQL from './sources/SQL';
 import { on, off } from './utils/events';
 import { isBrowserSupported, unsupportedBrowserReasons } from './renderer/Renderer';
@@ -68,7 +68,7 @@ import * as basemaps from './basemaps';
  */
 import { version } from '../package.json';
 
-const source = { Dataset, SQL, GeoJSON, MVT, BQMVT };
+const source = { Dataset, SQL, GeoJSON, MVT, BigQueryTileset };
 
 export { version, on, off, isBrowserSupported, unsupportedBrowserReasons, setDefaultAuth, setDefaultConfig, source, expressions, Layer, Viz, Interactivity, basemaps };
 export default { version, on, off, setDefaultAuth, setDefaultConfig, source, expressions, Layer, Viz, Interactivity, basemaps };
