+++ src/sources/Workers.worker.js
@@ -1,6 +1,6 @@
 import { MVTWorker } from './MVTWorker';
-import { BQMVTWorker } from './bigquery/BQMVTWorker';
 import { WindshaftWorker } from '../client/WindshaftWorker';
+import { BigQueryTilesetWorker } from './bigquery/BQTilesetWorker';
 
 // This file publishes a Web Worker onmessage function that will redirect incoming messages
 // based on an ID (`workerName`) defined by the message sender
@@ -8,8 +8,8 @@ import { WindshaftWorker } from '../client/WindshaftWorker';
 
 const workers = {
     MVT: new MVTWorker(),
-    BQMVT: new BQMVTWorker(),
-    windshaft: new WindshaftWorker()
+    windshaft: new WindshaftWorker(),
+    BigQueryTileset: new BigQueryTilesetWorker()
 };
 
 onmessage = function (event) {
