+++ src/sources/bigquery/BQTileClient.js
@@ -1,13 +1,13 @@
-import BQDataframeCache from './BQDataframeCache';
+import BigQueryTileCache from './BQTileCache';
 import { rTiles } from '../../client/rsys';
 import { isSetsEqual } from '../../utils/util';
 
-export default class BQTileClient {
+export default class BigQueryTileClient {
     constructor () {
         this._nextGroupID = 0;
         this._currentRequestGroupID = 0;
         this._oldDataframes = [];
-        this._cache = new BQDataframeCache();
+        this._cache = new BigQueryTileCache();
     }
 
     bindLayer (addDataframe) {
@@ -22,7 +22,7 @@ export default class BQTileClient {
 
     free () {
         this._cache.free();
-        this._cache = new BQDataframeCache();
+        this._cache = new BigQueryTileCache();
         this._oldDataframes = [];
     }
