+++ src/sources/bigquery/BQTilesetWorker.js
@@ -7,8 +7,7 @@ import DummyDataframe from '../../renderer/dataframe/DummyDataframe';
 import CartoValidationError, { CartoValidationErrorTypes } from '../../errors/carto-validation-error';
 import CartoRuntimeError, { CartoRuntimeErrorTypes } from '../../errors/carto-runtime-error';
 import { GEOMETRY_TYPE } from '../../utils/geometry';
-import BQClient from './BQClient';
-import { fetchTiles } from './utils';
+import BigQueryTilesetClient from './BQTilesetClient';
 
 // TODO import correctly
 const RTT_WIDTH = 1024;
@@ -22,7 +21,7 @@ const MVT_TO_CARTO_TYPES = {
     3: GEOMETRY_TYPE.POLYGON
 };
 
-export class BQMVTWorker {
+export class BigQueryTilesetWorker {
     // Worker API
     onmessage (event) {
         this.processEvent(event).then(postMessage);
@@ -34,7 +33,7 @@ export class BQMVTWorker {
             this.castMetadata(params.metadata);
             this.metadata = params.metadata;
         }
-        const dataframes = await this._requestDataframes(params.tiles, params.bqSource, params.layerID, params.tilesetMetadata, this.metadata);
+        const dataframes = await this._requestDataframes(params.tiles, params.tilesetData, params.tilesetMetadata, this.metadata);
         return {
             mID: params.mID,
             dataframes
@@ -46,15 +45,15 @@ export class BQMVTWorker {
         metadata.setCodecs();
     }
 
-    async _requestDataframes (tiles, bqSource, layerID, tilesetMetadata, metadata) {
+    async _requestDataframes (tiles, tilesetData, tilesetMetadata, metadata) {
         const dataframes = [];
-        const client = new BQClient(bqSource.project, bqSource.token);
-        const responseTiles = await fetchTiles(client, tiles, bqSource.dataset, bqSource.tileset, tilesetMetadata);
+        const client = new BigQueryTilesetClient(tilesetData.project, tilesetData.token);
+        const responseTiles = await client.fetchTiles(tiles, tilesetData.dataset, tilesetData.tileset, tilesetMetadata);
         for (let i = 0; i < tiles.length; i++) {
             const t = tiles[i];
             const responseTile = responseTiles && responseTiles.find((rt) => (rt.x === t.x && rt.y === t.y && rt.z === t.z));
             if (responseTile) {
-                const dataframe = await this.responseToDataframeTransformer(responseTile, layerID, metadata);
+                const dataframe = await this.responseToDataframeTransformer(responseTile, metadata);
                 dataframes.push(dataframe);
             } else {
                 dataframes.push({ x: t.x, y: t.y, z: t.z, empty: true });
@@ -63,21 +62,21 @@ export class BQMVTWorker {
         return dataframes;
     }
 
-    async responseToDataframeTransformer ({ x, y, z, buffer }, layerID, metadata) {
+    async responseToDataframeTransformer ({ x, y, z, buffer }, metadata) {
         const MVT_EXTENT = metadata.extent;
         if (buffer.byteLength === 0) {
             return { x, y, z, empty: true };
         }
         const tile = new VectorTile(new Protobuf(buffer));
 
-        if (Object.keys(tile.layers).length > 1 && !layerID) {
+        if (Object.keys(tile.layers).length > 1) {
             throw new CartoValidationError(
                 `LayerID parameter wasn't specified and the MVT tile contains multiple layers: ${JSON.stringify(Object.keys(tile.layers))}.`,
                 CartoValidationErrorTypes.MISSING_REQUIRED
             );
         }
 
-        const mvtLayer = tile.layers[layerID || Object.keys(tile.layers)[0]]; // FIXME this!!!
+        const mvtLayer = tile.layers[Object.keys(tile.layers)[0]];
 
         if (!mvtLayer) {
             return { empty: true };
